feat: Implementar navegação tabs nas  Visualizacao de Casos, Dashboard e Perfil
