+++ src/app/(tabs)/Dashboard.tsx
@@ -1,0 +1,17 @@
+import {StyleSheet, Text, View} from 'react-native'
+
+export default function Dashboard(){
+    return (
+        <View>
+            <Text>Página Dashboard</Text>
+        </View>
+    )
+}
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        alignItems: 'center',
+        backgroundColor: 'white',
+    },
+})
