+++ src/app/(tabs)/Perfil.tsx
@@ -1,0 +1,17 @@
+import {StyleSheet, Text, View} from 'react-native'
+
+export default function Perfil(){
+    return (
+        <View>
+            <Text>Página Perfil</Text>
+        </View>
+    )
+}
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        alignItems: 'center',
+        backgroundColor: 'white',
+    },
+})
