+++ src/app/(tabs)/VisualizacaoCasos.tsx
@@ -1,0 +1,17 @@
+import {StyleSheet, Text, View} from 'react-native'
+
+export default function VisualizacaoCasos(){
+    return (
+        <View>
+            <Text>Home</Text>
+        </View>
+    )
+}
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        alignItems: 'center',
+        backgroundColor: 'white',
+    },
+})
